COT/requirements.txt
torchinfo
matplotlib
mmcv-full
scikit-learn
h5py
